+++ src/lib/config.ts
@@ -61,8 +61,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.2",
-  sdkVersion: "0.1.0-alpha.9",
-  genVersion: "2.463.1",
+  sdkVersion: "0.1.0-alpha.10",
+  genVersion: "2.467.4",
   userAgent:
-    "speakeasy-sdk/typescript 0.1.0-alpha.9 2.463.1 0.0.2 syllable-sdk",
+    "speakeasy-sdk/typescript 0.1.0-alpha.10 2.467.4 0.0.2 syllable-sdk",
 } as const;