+++ src/sdk/sdk.ts
@@ -8,12 +8,14 @@ import { Channels } from "./channels.js";
 import { Conversations } from "./conversations.js";
 import { CustomMessages } from "./custommessages.js";
 import { Dashboards } from "./dashboards.js";
+import { Directory } from "./directory.js";
 import { Events } from "./events.js";
 import { Prompts } from "./prompts.js";
 import { Services } from "./services.js";
 import { SessionLabels } from "./sessionlabels.js";
 import { Sessions } from "./sessions.js";
 import { Tools } from "./tools.js";
+import { V1 } from "./v1.js";
 
 export class SyllableSDK extends ClientSDK {
   private _agents?: Agents;
@@ -69,5 +71,15 @@ export class SyllableSDK extends ClientSDK {
   private _tools?: Tools;
   get tools(): Tools {
     return (this._tools ??= new Tools(this._options));
+  }
+
+  private _directory?: Directory;
+  get directory(): Directory {
+    return (this._directory ??= new Directory(this._options));
+  }
+
+  private _v1?: V1;
+  get v1(): V1 {
+    return (this._v1 ??= new V1(this._options));
   }
 }